+++ Ambassador/src/utopia/ambassador/rest/resource/service/auth/AuthNode.scala
@@ -15,6 +15,7 @@ import utopia.citadel.util.CitadelContext._
 import utopia.exodus.rest.util.AuthorizedContext
 import utopia.exodus.util.ExodusContext.handleError
 import utopia.exodus.util.ExodusContext.uuidGenerator
+import utopia.flow.collection.LazyVector
 import utopia.flow.datastructure.template.MapLike
 import utopia.flow.time.Now
 import utopia.flow.util.CollectionExtensions._
@@ -36,7 +37,8 @@ class AuthNode(target: ServiceTarget, tokenAcquirer: AcquireTokens, redirectors:
 	
 	override val name = "auth"
 	
-	override lazy val children = Vector(new AuthPreparationNode(target), new AuthResponseNode(target, tokenAcquirer))
+	override val children = LazyVector.fromFunctions(
+		() => new AuthPreparationNode(target), () => new AuthResponseNode(target, tokenAcquirer))
 	
 	
 	// IMPLEMENTED  -----------------------------